Added fs ids.
