+++ tests/test_server.c
@@ -1,4 +1,4 @@
-#include "net.h"
+#include "net/server.h"
 
 int test_server(char *port)
 {
