+++ tools/fs_hex_decoder.py
@@ -1,5 +1,4 @@
-hex_str = "A81A817667267262767726666A0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005542233222776A4442233222666A4544233222444227672677A444423322266626AA4444522233225666226A44452222332266224277726666626662267767667626A4444544233224277726666626666766AA5455445233222766662666622255422266662222622A4444455233222766662666622255422266662222622AA55454233222776762676662A554545545544523322276666266662225545422266662222622A554545444445523322276666266662225545422266662222622AA4444544544552332242255422242A2244445445445523322422554542224225542A767322254554452222444445522225545454444455222244442A9666672272224442A96662224444522262224544222244445445445522225455445222244442222554545545544522224445222244445442AA222222225455522222222AA76773A96666266666A96666222222222222222222222666767727622266627672677A96666276775666622222222222666767727622266627677727677566662677A9666627677576766667622222266676772762226662767772767757676666762677AA767756666322276777276775666626222545544522224444455222255454544444552A96666722722244422767772A966622244445222622244422767772242677222444454454455222254554452223222554545545544522224445222244445442AA7677576766667632227677727677576766667626222545544522224444455222255454544444552A96666722722244422767772A966622244445222622244422767772242677222444454454455222254554452223222554545545544522224445222244445442AA7677576776732227677727677576776726222545544522224444455222255454544444552A96666722722244422767772A966622244445222622244422767772242677222444454454455222254554452223222554545545544522224445222244445442AA7677566666732227677727677566666726222545544522224444455222255454544444552A96666722722244422767772A966622244445222622244422767772242677222444454454455222254554452223222554545545544522224445222244445442AA666663A9762276266622A0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000014A2A1AC76776666277711A1A84666666618D1C1A0A0A0A0A0A0A0A0A0A0A0A0A0A0A012"
+hex_str = "0A83100A801074686973206973206120746573742066696C650A000000000000000000000000000000000000000000000000000000000000617765736F6D652066696C65210A0A6E65776C696E653F213F213F0A0A00000000000000000000000000000000000000000053524320203A3D202E2F7372630A42494E20203A3D202E2F62696E0A45584543203A3D20242842494E292F7466732E6F75740A4D41494E203A3D202E2F636C692E630A0A464C4147532020203A3D202D57616C6C202D670A4C494253202020203A3D202D6C6D202D4C2F7573722F6C6F63616C2F6C6962202D6C70726F746F6275662D630A494E434C554445203A3D202D492F7573722F6C6F63616C2F696E636C7564650A0A534F5552434553203A3D2024287368656C6C2066696E6420242853524329202D6E616D6520272A2E6327290A48454144455253203A3D2024287368656C6C2066696E6420242853524329202D6E616D6520272A2E6827290A0A50524F544F203A3D202E2F70726F746F2F6275696C642F0A50524F544F5F534F5552434553203A3D2024287368656C6C2066696E6420242850524F544F29202D6E616D6520272A2E6327290A50524F544F5F48454144455253203A3D2024287368656C6C2066696E6420242850524F544F29202D6E616D6520272A2E6827290A0A494E434C5544455F44495253203A3D202D49242853524329202D492E0A2320494E434C5544455F44495253203A3D202D49242850524F544F29202D492428535243290A7466733A202428534F555243455329202428484541444552532920242850524F544F5F48454144455253292024284D41494E290A096D6B646972202D7020242842494E290A09676363202428464C41475329202D6F2024284558454329202428494E434C5544455F4449525329202428534F5552434553292024284D41494E2920242850524F544F5F534F5552434553292024284C49425329202428494E434C554445290A0A23202D2D2D2D2D205445535453202D2D2D2D2D20230A0A746573743A0A096D616B6520636C65616E0A096D616B65202020202020202020202020202020202626202E2F6D656D746573742E7368202E2F62696E2F7466732E6F75740A096D616B6520746573745F6D61696E2020202020202626202E2F6D656D746573742E7368202E2F62696E2F74657374732F746573745F6D61696E2E6F75740A096D616B6520746573745F73657269616C697A65202626202E2F6D656D746573742E7368202E2F62696E2F74657374732F746573745F73657269616C697A652E6F75740A0A746573745F6D61696E3A202E2F74657374732F746573745F6D61696E2E63202428534F555243455329202428484541444552532920242850524F544F5F48454144455253290A096D6B646972202D7020242842494E292F74657374732F0A09676363202428464C41475329202D6F20242842494E292F74657374732F24402E6F7574202428494E434C5544455F4449525329202428534F55524345532920243C20242850524F544F5F534F5552434553292024284C49425329202428494E434C554445290A0A746573745F73657269616C697A653A202E2F74657374732F746573745F73657269616C697A652E63202428534F555243455329202428484541444552532920242850524F544F5F48454144455253290A096D6B646972202D7020242842494E292F74657374732F0A09676363202428464C41475329202D6F20242842494E292F74657374732F24402E6F7574202428494E434C5544455F4449525329202428534F55524345532920243C20242850524F544F5F534F5552434553292024284C49425329202428494E434C554445290A0A746573745F7365727665723A202E2F74657374732F746573745F7365727665722E63202428534F555243455329202428484541444552532920242850524F544F5F48454144455253290A096D6B646972202D7020242842494E292F74657374732F0A09676363202428464C41475329202D6F20242842494E292F74657374732F24402E6F7574202428494E434C5544455F4449525329202428534F55524345532920243C20242850524F544F5F534F5552434553292024284C49425329202428494E434C554445290A0A746573745F636C69656E743A202E2F74657374732F746573745F636C69656E742E63202428534F555243455329202428484541444552532920242850524F544F5F48454144455253290A096D6B646972202D7020242842494E292F74657374732F0A09676363202428464C41475329202D6F20242842494E292F74657374732F24402E6F7574202428494E434C5544455F4449525329202428534F55524345532920243C20242850524F544F5F534F5552434553292024284C49425329202428494E434C554445290A0A636C65616E3A0A09726D202D72662062696E2F2A0A000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000125A0A380A100A0C7465737466696C652E74787410140A130A0D7465737466696C65322E747874101D18320A0F0A084D616B6566696C6510880D18640A000A000A000A000A000A000A000A000A000A000A000A000A000A000A001003"
 s = "".join([chr(int(hex_str[i:i+2], 16)) for i in range(0, len(hex_str), 2)])
 print(s)
-
